requirements.txt
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
pandas==1.5.3
requests>=2.32.2
python-dotenv==0.21.0
langchain-core>=0.1.16,<0.2
urllib3>=2